Don't start Redis if the password is missing
Without error handling the redis server will be started with a required
password of empty string, which has the same effects as not requiring a
password.

This can happen both if the password is not defined in Vault, or if the
password contains the character '&' and is not wrapped in quotes.